solving helper
In this version, I added a software module used for teaching users to
solve Rubik's Cube.
     The module is a typical expert system. Knowledge Base storage
pattern and solving rules. By building pattern analysis tree, the
inference engine to match the corresponding pattern to apply the
solving rules. At the same time, in this version I add the pattern and
solving rules editor. Pattern as well as solving rules recorded in the
Knowledge Base as the form of a number string.