20120115
In this version, I separated State from Pattern.
The state of Rubik's cube shows the current solution degree.
And the patterns which are in memory now are  based on the state
instead of all in memory.
I change the data model by adding an array of six elements which shows
the difference between logic orientation and MC's orientation.